This is the concusion, Intro and abstract
Documented in past tense with use of present / present continuous
Introduction containd (a) a definition of the problem (b) motivation behind the research (c) a definition of the research objectives (d) a section describing the various chapters in your study.
Conclusion contains (a) summary of the findings (b) recommendations for future research (c) an evaluation of the findings outlining the strengths/weaknesses of the research
Abstract contains a summary of the dissertation with highlights from the study.  (500 words length)
References provided in context
Passive style of writing (no first person)
Tables and Figures were not added at this stage